output JSON from hillclimber
